+++ actions/adyen/actions.config.yaml
@@ -1,0 +1,9 @@
+start-session:
+  function: ./start-session/index.js
+  web: 'yes'
+  runtime: nodejs:20
+  inputs:
+    LOG_LEVEL: debug
+  annotations:
+    require-adobe-auth: true
+    final: true
