+++ app.config.yaml
@@ -20,6 +20,10 @@ application:
       #    license: Apache-2.0
       #    actions:
       #      $include: ./actions/ingestion/actions.config.yaml
+      adyen:
+        license: Apache-2.0
+        actions:
+          $include: ./actions/adyen/actions.config.yaml
       webhook:
         license: Apache-2.0
         actions:
