--- test/actions/customer/external/created/sender.test.js
@@ -1,33 +1,0 @@
-/*
-Copyright 2022 Adobe. All rights reserved.
-This file is licensed to you under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License. You may obtain a copy
-of the License at http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software distributed under
-the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
-OF ANY KIND, either express or implied. See the License for the specific language
-governing permissions and limitations under the License.
-*/
-
-jest.mock('../../../../../actions/customer/commerce-customer-api-client')
-const { createCustomer } = require('../../../../../actions/customer/commerce-customer-api-client')
-
-const sender = require('../../../../../actions/customer/external/created/sender')
-
-describe('Given customer external created sender', () => {
-  describe('When method sendData is defined', () => {
-    test('Then is an instance of Function', () => {
-      expect(sender.sendData).toBeInstanceOf(Function)
-    })
-  })
-  describe('When method sendData is called', () => {
-    test('Then create customer is called', async () => {
-      const params = {}
-      const transformed = {}
-      const preprocess = {}
-      await sender.sendData(params, transformed, preprocess)
-      expect(createCustomer).toHaveBeenCalled()
-    })
-  })
-})
